change of Voting5
数据类型修改，voting里的规划中 IloExpr -> IloIntVar ，
flow.h中evaluate的修改

+++ Voting5/LP.h
@@ -86,7 +86,7 @@ double LP(PGraph *g,vector<Req*> &reqL)
 		}
 		for(int d=0;d<K;d++)
 		{
-			int temp_bw=Inf;
+			double temp_bw=Inf;
 			for(int i=0;i<g->m;i++)
 			{
 				if(solver.getValue(x[d][i])>0)
